separated materials adapter
